added dummy for order

+++ src/main/java/com/se330/coffee_shop_management_backend/dto/request/product/ProductVariantCreateRequestDTO.java
@@ -3,6 +3,7 @@ package com.se330.coffee_shop_management_backend.dto.request.product;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -10,6 +11,6 @@ import java.util.UUID;
 public class ProductVariantCreateRequestDTO {
     private String variantTierIdx;
     private Boolean variantDefault;
-    private Long variantPrice;
+    private BigDecimal variantPrice;
     private UUID product;
 }
