+++ src/main/java/com/se330/coffee_shop_management_backend/dto/request/product/ProductVariantUpdateRequestDTO.java
@@ -3,6 +3,7 @@ package com.se330.coffee_shop_management_backend.dto.request.product;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -13,7 +14,7 @@ public class ProductVariantUpdateRequestDTO {
     private Boolean variantDefault;
     private String variantSlug;
     private int variantSort;
-    private Long variantPrice;
+    private BigDecimal variantPrice;
     private Boolean variantIsPublished;
     private Boolean variantIsDeleted;
     private UUID product;
